pyproject.toml
[project]
name = "bandit-exploration"
version = "0.1.0"
description = "My personal reimplementation of various bandit algorithms."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cvxpy>=1.7.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "ray>=2.48.0",
    "tqdm>=4.67.1",
    "cyclopts>=3.0.0,<4.0.0",
    "loguru>=0.7.0,<0.8.0",
    "pyqt5<5.16",
    "pyqtwebengine<5.16",
    "loguru<1.0",
    "icecream>=2.0.0,<3.0.0",
    "ipykernel>=6.30.1",
    "polars>=1.32.2",
    "ecos>=2.0.14",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
    "scipy-stubs>=1.16.1.0",
    "types-tqdm>=4.67.0.20250809",
    "viztracer>=1.0.4",
]
